refactor(types): centralize types and refactor navigation paths
- Move types to dedicated types/types.ts file
- Update dummy data to import from new types location
- Refactor navigation paths to use enum in paths.ts
- Update AuthNavigator, LoginScreen, and other navigation references to use Paths enum
- Update navigation types to include new paths
- Update domain index to export post hooks module
- Improve navigation type safety with enum-based paths
