+++ react-native/Kryos/src/hooks/domain/index.ts
@@ -1,1 +1,2 @@
+export * from './post';
 export { useUser } from './user/useUser';
