+++ react-native/Kryos/src/navigation/AuthNavigator.tsx
@@ -6,14 +6,16 @@ import LoginScreen from '@/screens/Auth/LoginScreen';
 import SignupScreen from '@/screens/Auth/SignupScreen';
 import FeedScreen from '@/screens/Feed/FeedScreen';
 
+import { Paths } from './paths';
+
 const Stack = createStackNavigator<AuthStackParamList>();
 
 function AuthNavigator() {
   return (
-    <Stack.Navigator initialRouteName="Feed" screenOptions={{ headerShown: false }}>
-      <Stack.Screen component={FeedScreen} name="Feed" />
-      <Stack.Screen component={LoginScreen} name="Login" />
-      <Stack.Screen component={SignupScreen} name="Signup" />
+    <Stack.Navigator initialRouteName={Paths.Feed} screenOptions={{ headerShown: false }}>
+      <Stack.Screen component={FeedScreen} name={Paths.Feed} />
+      <Stack.Screen component={LoginScreen} name={Paths.Login} />
+      <Stack.Screen component={SignupScreen} name={Paths.Signup} />
     </Stack.Navigator>
   );
 }
